+++ modules/time.js
@@ -1,0 +1,5 @@
+export let timing = () => {
+    const timeDate = document.getElementById('dated')
+    const date = new Date();
+    timeDate.innerHTML = date;
+  }
